test: metamask balance updates after actions
