+++ packages/app/playwright/commons/index.ts
@@ -3,3 +3,4 @@ export * from './button';
 export * from './visit';
 export * from './text';
 export * from './seedWallet';
+export * from './wallet';
